Week 1 version 1 of to do list:
Your task is to design the data layer for a to-do list app. For inspiration, here's 40 examples. Create the core data structures that track state, along with an API that can be connected to any frontend (command-line, web, desktop GUI, etc).

Initial Design
In the basic version, a task is either done or not done. A user should be able to create and delete todo items, view the list to do, and mark items as done.